requirements.txt
selene>=2.0.0rc7
webdriver-manager==4.0.2
pytest==8.3.2
selenium~=4.25.0
pypdf==5.0.1
openpyxl==3.1.5
allure-pytest==2.13.5
requests==2.32.3
allure-pytest==2.13.5